+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased][unreleased]
 
+(no unreleased changes yet)
+
+
+## [v0.11.0][]
+
+**Release date: 2020-01-28**
+
 ### Added
 
 * `penman.model.Model.check()` ([#55][])
@@ -566,6 +573,7 @@ First release with very basic functionality.
 [v0.9.0]: ../../releases/tag/v0.9.0
 [v0.9.1]: ../../releases/tag/v0.9.1
 [v0.10.0]: ../../releases/tag/v0.10.0
+[v0.11.0]: ../../releases/tag/v0.11.0
 [README]: README.md
 
 [#4]: https://github.com/goodmami/penman/issues/4